+++ yaralyzer/bytes_match.py
@@ -71,11 +71,19 @@ class BytesMatch:
             highlight_style: str = YaralyzerConfig.HIGHLIGHT_STYLE
         ) -> 'BytesMatch':
         """Build a BytesMatch from a yara string match. matched_against is the set of bytes yara was run against"""
+        # Don't duplicate the labeling if rule_name and yara_str are the same
+        pattern_label = yara_str[1]
+
+        if pattern_label == '$' + rule_name:
+            label = pattern_label
+        else:
+            label = rule_name + ': ' + pattern_label
+
         return cls(
             matched_against=matched_against,
             start_idx=yara_str[0],
             length=len(yara_str[2]),
-            label=rule_name + ': ' + yara_str[1],
+            label=label,
             ordinal=ordinal,
             highlight_style=highlight_style)