+++ yaralyzer/helpers/bytes_helper.py
@@ -62,7 +62,7 @@ def rich_text_view_of_raw_bytes(_bytes: bytes, bytes_match: BytesMatch) -> Text:
 
 
 def hex_view_of_raw_bytes(_bytes: bytes, bytes_match: BytesMatch) -> Text:
-    hex_str = hex_string(_bytes)
+    hex_str = hex_text(_bytes)
     highlight_start_idx = bytes_match.highlight_start_idx * 3
     highlight_end_idx = bytes_match.highlight_end_idx * 3
     hex_str.stylize(bytes_match.highlight_style, highlight_start_idx, highlight_end_idx)
@@ -70,8 +70,12 @@ def hex_view_of_raw_bytes(_bytes: bytes, bytes_match: BytesMatch) -> Text:
     return Text("\n").join([Text('  ').join(line.wrap(console, HEX_CHARS_PER_GROUP * 3)) for line in lines])
 
 
-def hex_string(_bytes: bytes) -> Text:
-    return Text(' '.join([hex(b).removeprefix('0x').rjust(2, '0') for i, b in enumerate(_bytes)]), style=GREY)
+def hex_text(_bytes: bytes) -> Text:
+    return Text(hex_string(_bytes), style=GREY)
+
+
+def hex_string(_bytes: bytes) -> str:
+    return ' '.join([hex(b).removeprefix('0x').rjust(2, '0') for i, b in enumerate(_bytes)])
 
 
 def print_bytes(bytes_array: bytes, style=None) -> None: