+++ yaralyzer/helpers/file_helper.py
@@ -10,8 +10,8 @@ def timestamp_for_filename() -> str:
 
 def files_in_dir(dir: str, with_extname: Optional[str] = None) -> List[str]:
     """paths for non dot files, optionally ending in 'with_extname'"""
-    files = [path.join(dir, file) for file in listdir(dir)]
-    files = [file for file in files if not (path.isdir(file) or file.startswith('.'))]
+    files = [path.join(dir, file) for file in listdir(dir) if not file.startswith('.')]
+    files = [file for file in files if not path.isdir(file)]
 
     if with_extname:
         return files_with_extname(files, with_extname)