+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # NEXT RELEASE
+* Add `--hex-pattern` command line option
+* Add `--patterns-label` command line option
 
 # 0.3.3
 * Refactor `file_export` and `rich_console`